feat: Add client and server app classes
This commit adds the `ClientApp` and `ServerApp` classes to the `net_enforcement` module. These classes are responsible for running the client and server applications respectively. They are initialized with a `Console` object from the `rich` library and have a method `run_client` and `run_server` respectively, which are currently empty.

The purpose of this commit is to lay the foundation for the client and server functionality in the `net_enforcement` module.